backend/requirements.txt
# ProManage Backend Requirements
# Generated from pyproject.toml for Passenger deployment

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.6
authlib>=1.3.0
itsdangerous>=2.1.2
pymupdf>=1.26.7
sqlalchemy>=2.0.45
pymysql>=1.1.2
psycopg2-binary>=2.9.9
email-validator>=2.3.0
cryptography>=42.0.0
requests>=2.32.5
stripe>=7.0.0
pyzbar>=0.1.9
pillow>=12.0.0

# Optional: ASGI to WSGI adapter if Passenger doesn't support ASGI
# a2wsgi>=1.10.0
